.claude/mcp/requirements.txt
fastmcp>=2.0.0,<3.0.0
sentence-transformers>=3.0.0,<4.0.0
faiss-cpu>=1.8.0,<2.0.0
rank-bm25>=0.2.2,<1.0.0
numpy>=1.26.0,<3.0.0
